Updated cargo deps/rust version, fixed sqlite migrations and added album page

+++ src/models/mod.rs
@@ -147,8 +147,8 @@ impl Responder for MessageResponse {
     ApplicationPage = Page<ApplicationData>
 )]
 pub struct Page<T> {
-    pub page: usize,
-    pub pages: usize,
+    pub page: u64,
+    pub pages: u64,
     pub items: Vec<T>,
 }
