+++ src/routes/user/application.rs
@@ -28,7 +28,7 @@ pub fn get_routes() -> Scope {
 #[get("/{page_number}")]
 async fn list(
     service: web::Data<ApplicationService>,
-    params: web::Path<(String, usize)>,
+    params: web::Path<(String, u64)>,
     user: Auth<auth_role::User>,
 ) -> impl Responder {
     let (user_id, page_number) = params.to_owned();
